+++ GridKit/Model/PhasorDynamics/BusBase.hpp
@@ -62,17 +62,6 @@ namespace GridKit
         // No time to update in bus models
       }
 
-      virtual void setTolerances(RealT& rtol, RealT& atol) const override
-      {
-        rtol = rtol_;
-        atol = atol_;
-      }
-
-      virtual void setMaxSteps(IdxT& msa) const override
-      {
-        msa = max_steps_;
-      }
-
       virtual ScalarT&       Vr()       = 0;
       virtual const ScalarT& Vr() const = 0;
       virtual ScalarT&       Vi()       = 0;
@@ -110,6 +99,16 @@ namespace GridKit
       const std::vector<bool>& tag() const override
       {
         return tag_;
+      }
+
+      std::vector<ScalarT>& absoluteTolerance() override
+      {
+        return abs_tol_;
+      }
+
+      const std::vector<ScalarT>& absoluteTolerance() const override
+      {
+        return abs_tol_;
       }
 
       MatrixT& getJacobian() override
@@ -177,6 +176,7 @@ namespace GridKit
       std::vector<ScalarT> y_;
       std::vector<ScalarT> yp_;
       std::vector<bool>    tag_;
+      std::vector<ScalarT> abs_tol_;
       std::vector<ScalarT> f_;
 
       MatrixT J_;