+++ GridKit/Model/PhasorDynamics/Exciter/IEEET1/Ieeet1.hpp
@@ -72,6 +72,7 @@ namespace GridKit
         using Component<ScalarT, IdxT>::nnz_;
         using Component<ScalarT, IdxT>::size_;
         using Component<ScalarT, IdxT>::tag_;
+        using Component<ScalarT, IdxT>::abs_tol_;
         using Component<ScalarT, IdxT>::time_;
         using Component<ScalarT, IdxT>::y_;
         using Component<ScalarT, IdxT>::yp_;
@@ -104,6 +105,7 @@ namespace GridKit
         int verify() const override final;
         int initialize() override final;
         int tagDifferentiable() override final;
+        int setAbsoluteTolerance(RealT rel_tol) override final;
         int evaluateResidual() override final;
         int evaluateJacobian() override final;