+++ GridKit/Model/PhasorDynamics/SignalNode/SignalNode.hpp
@@ -37,6 +37,7 @@ namespace GridKit
       virtual int allocate() override final;
       virtual int initialize() override final;
       virtual int tagDifferentiable() override final;
+      virtual int setAbsoluteTolerance(RealT) override final;
       virtual int evaluateResidual() override final;
       virtual int evaluateJacobian() override final;
 
@@ -70,17 +71,6 @@ namespace GridKit
         // No time to update in signal nodes
       }
 
-      virtual void setTolerances(RealT& rtol, RealT& atol) const override final
-      {
-        rtol = rtol_;
-        atol = atol_;
-      }
-
-      virtual void setMaxSteps(IdxT& msa) const override final
-      {
-        msa = max_steps_;
-      }
-
       std::vector<ScalarT>& y() override final
       {
         return y_;
@@ -109,6 +99,16 @@ namespace GridKit
       const std::vector<bool>& tag() const override final
       {
         return tag_;
+      }
+
+      std::vector<ScalarT>& absoluteTolerance() override
+      {
+        return abs_tol_;
+      }
+
+      const std::vector<ScalarT>& absoluteTolerance() const override
+      {
+        return abs_tol_;
       }
 
       MatrixT& getJacobian() override final
@@ -140,6 +140,7 @@ namespace GridKit
       std::vector<ScalarT> y_;
       std::vector<ScalarT> yp_;
       std::vector<bool>    tag_;
+      std::vector<ScalarT> abs_tol_;
       std::vector<ScalarT> f_;
 
       MatrixT J_;