+++ GridKit/Model/PhasorDynamics/SignalNode/SignalNodeImpl.hpp
@@ -38,6 +38,24 @@ namespace GridKit
     {
       return 0;
     }
+    
+    /**
+     * @brief Compute the absolute tolerance for each variable in the model
+     * 
+     * @param rel_tol The relative tolerance which can be used to pick the
+     *        absolute tolerance.
+     * @tparam ScalarT Scalar data type
+     * @tparam IdxT Index data type
+     * @return int 0 if successful, non-zero otherwise.
+     * 
+     * This represents a "noise" level close to zero for which pure relative
+     * error cannot be used.
+     */
+    template <class ScalarT, typename IdxT>
+    int SignalNode<ScalarT, IdxT>::setAbsoluteTolerance(RealT)
+    {
+      return 0;
+    }
 
     template <class ScalarT, typename IdxT>
     int SignalNode<ScalarT, IdxT>::evaluateResidual()