+++ GridKit/Model/PhasorDynamics/SynchronousMachine/GENROUwS/Genrou.hpp
@@ -75,6 +75,7 @@ namespace GridKit
       using Component<ScalarT, IdxT>::nnz_;
       using Component<ScalarT, IdxT>::size_;
       using Component<ScalarT, IdxT>::tag_;
+      using Component<ScalarT, IdxT>::abs_tol_;
       using Component<ScalarT, IdxT>::time_;
       using Component<ScalarT, IdxT>::y_;
       using Component<ScalarT, IdxT>::yp_;
@@ -133,6 +134,7 @@ namespace GridKit
       int verify() const override final;
       int initialize() override final;
       int tagDifferentiable() override final;
+      int setAbsoluteTolerance(RealT) override final;
       int evaluateResidual() override final;
 
       // Still to be implemented