+++ GridKit/Model/PhasorDynamics/SynchronousMachine/GenClassical/GenClassical.hpp
@@ -39,6 +39,7 @@ namespace GridKit
       using Component<ScalarT, IdxT>::nnz_;
       using Component<ScalarT, IdxT>::size_;
       using Component<ScalarT, IdxT>::tag_;
+      using Component<ScalarT, IdxT>::abs_tol_;
       using Component<ScalarT, IdxT>::time_;
       using Component<ScalarT, IdxT>::y_;
       using Component<ScalarT, IdxT>::yp_;
@@ -74,6 +75,7 @@ namespace GridKit
       int allocate() override final;
       int initialize() override final;
       int tagDifferentiable() override final;
+      int setAbsoluteTolerance(RealT) override;
       int evaluateResidual() override final;
 
       int verify() const override final