+++ GridKit/Model/PowerElectronics/Inductor/Inductor.cpp
@@ -41,6 +41,8 @@ namespace GridKit
     y_.resize(static_cast<size_t>(size_));
     yp_.resize(static_cast<size_t>(size_));
     f_.resize(static_cast<size_t>(size_));
+    tag_.resize(static_cast<size_t>(size_));
+    abs_tol_.resize(static_cast<size_t>(size_));
 
     return 0;
   }
@@ -60,6 +62,25 @@ namespace GridKit
   template <class ScalarT, typename IdxT>
   int Inductor<ScalarT, IdxT>::tagDifferentiable()
   {
+    return 0;
+  }
+
+  /**
+   * @brief Compute the absolute tolerance for each variable in the model
+   * 
+   * @param rel_tol The relative tolerance which can be used to pick the
+   *        absolute tolerance.
+   * @tparam ScalarT Scalar data type
+   * @tparam IdxT Index data type
+   * @return int 0 if successful, non-zero otherwise.
+   * 
+   * This represents a "noise" level close to zero for which pure relative
+   * error cannot be used.
+   */
+  template <class ScalarT, typename IdxT>
+  int Inductor<ScalarT, IdxT>::setAbsoluteTolerance(RealT rel_tol)
+  {
+    std::fill(abs_tol_.begin(), abs_tol_.end(), rel_tol);
     return 0;
   }