+++ GridKit/Model/PowerFlow/Generator4Governor/Generator4Governor.hpp
@@ -26,6 +26,7 @@ namespace GridKit
     using ModelEvaluatorImpl<ScalarT, IdxT>::y_;
     using ModelEvaluatorImpl<ScalarT, IdxT>::yp_;
     using ModelEvaluatorImpl<ScalarT, IdxT>::tag_;
+    using ModelEvaluatorImpl<ScalarT, IdxT>::abs_tol_;
     using ModelEvaluatorImpl<ScalarT, IdxT>::f_;
     using ModelEvaluatorImpl<ScalarT, IdxT>::g_;
     using ModelEvaluatorImpl<ScalarT, IdxT>::yB_;